+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 20.01.2018
 
-* TODO: Describe initial release.
+* Added Pusher connector. Implemented by @abrahamSN
+* Fixes issue with socket-io private channels
+* Cleaned up
+
+## [0.0.1] - 20.01.2018
+
+* Initial release.
+* Added null connector
+* Added socketio connector